function/requirements.txt
azure-functions
psycopg2-binary
sendgrid==6.3.0